move istioctl out of pilot subdirectory
istioctl is cli for all of Istio, not solely Pilot. Move istioctl to
the top-level in istio/istio so that code reflects reality. This will
also make it easier to manage codeowners for cli specific issues
unrelated to Pilot and networking.
