+++ istioctl/OWNERS
@@ -1,0 +1,62 @@
+reviewers:
+  - andraxylia
+  - ayj
+  - GregHanson
+  - gyliu513
+  - ijsnellf
+  - kyessenov
+  - rshriram
+  - sebastienvas
+  - nmittler
+  - costinm
+  - zackbutcher
+  - ldemailly
+  - myidpt
+  - mandarjog
+  - jmuk
+  - lizan
+  - smawson
+  - frankbu
+  - esnible
+  - cmluciano
+  - louiscryan
+  - crazytan
+  - rkpagadala
+  - guptasu
+  - qiwzhang
+  - elevran
+  - liamawhite
+  - wenchenglu
+  - vadimeisenbergibm
+  - gnirodi
+approvers:
+  - andraxylia
+  - ayj
+  - GregHanson
+  - gyliu513
+  - ijsnellf
+  - kyessenov
+  - rshriram
+  - nmittler
+  - costinm
+  - zackbutcher
+  - ldemailly
+  - myidpt
+  - mandarjog
+  - jmuk
+  - lizan
+  - smawson
+  - frankbu
+  - esnible
+  - cmluciano
+  - louiscryan
+  - crazytan
+  - rkpagadala
+  - guptasu
+  - qiwzhang
+  - elevran
+  - liamawhite
+  - wenchenglu
+  - vadimeisenbergibm
+  - gnirodi
+
