+++ istioctl/deregister.go
