+++ istioctl/gendeployment/cmd_test.go
