+++ istioctl/gendeployment/testdata/default-values.yaml.golden
