+++ istioctl/main.go
@@ -42,8 +42,8 @@ import (
 	"github.com/spf13/cobra/doc"
 	"k8s.io/client-go/util/homedir"
 
+	"istio.io/istio/istioctl/gendeployment"
 	"istio.io/istio/pilot/cmd"
-	"istio.io/istio/pilot/cmd/istioctl/gendeployment"
 	"istio.io/istio/pilot/pkg/config/kube/crd"
 	"istio.io/istio/pilot/pkg/model"
 	"istio.io/istio/pilot/pkg/serviceregistry/kube"
